requirements.txt
aiohttp
aria2p
async_generator
bs4
beautifulsoup4>=4.8.1
cowpy
cryptg
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans>=2.4.0
google_images_download>=2.7.1
gTTS>=2.1.1
gTTS-token>=1.1.3
hachoir
httplib2
humanize
heroku3
lxml==4.5.2
git+https://github.com/johnwmillr/LyricsGenius.git
lyricsgenius
oauth2client
psycopg2
pybase64>=0.4.0
pylast==3.0.0
pySmartDL
python-barcode
python-dotenv
youtube-dl
pytz
qrcode
requests>=2.18.4
git+https://github.com/bisoncorps/search-engine-parser
speedtest-cli>=2.0.2
sqlalchemy>=1.2
spotdl
selenium>=3.141.0
telethon>=1.10.6
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse
spotify_token
twitter_scraper
chromedriver-autoinstaller
numpy
Pillow
telegraph
lottie
git+https://github.com/Joeclinton1/google-images-download.git
eksipy
pydub
emoji-country-flag
cairosvg